Fix #1: Renamed package name
