+++ .idea/dictionaries/trf.xml
@@ -1,0 +1,8 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="chinm">
+    <words>
+      <w>netra</w>
+      <w>vitpunerobotics</w>
+    </words>
+  </dictionary>
+</component>
