+++ app/src/main/res/layout/activity_color__detector.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Color_Detector">
+    tools:context="com.vitpunerobotics.netra.Color_Detector">
 
 
     <org.opencv.android.JavaCameraView
