+++ app/src/main/res/layout/activity_gesture.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Gesture">
+    tools:context="com.vitpunerobotics.netra.Gesture">
 
 </android.support.constraint.ConstraintLayout>
