+++ app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/yellow"
-    tools:context=".Home">
+    tools:context="com.vitpunerobotics.netra.Home">
 
     <ImageView
         android:id="@+id/imageView4"
