+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:background="@color/yellow"
     android:gravity="center_horizontal"
     android:orientation="vertical"
-    tools:context=".MainActivity">
+    tools:context="com.vitpunerobotics.netra.MainActivity">
 
 
     <ProgressBar
