+++ app/src/main/res/layout/activity_main2.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Main2Activity">
+    tools:context="com.vitpunerobotics.netra.Main2Activity">
 
 
     <org.opencv.android.JavaCameraView
