+++ app/src/main/res/layout/activity_main3.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Main3Activity">
+    tools:context="com.vitpunerobotics.netra.Main3Activity">
 
 
     <android.support.v7.widget.Toolbar
