+++ app/src/main/res/layout/activity_manual.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_weight="3"
     android:background="@color/yellow"
-    tools:context=".MainActivity">
+    tools:context="com.vitpunerobotics.netra.MainActivity">
 
     <LinearLayout
