+++ app/src/main/res/layout/activity_move_dist.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Move_dist">
+    tools:context="com.vitpunerobotics.netra.Move_dist">
 
     <org.opencv.android.JavaCameraView
         android:id="@+id/javaCameraView"
