+++ app/src/main/res/layout/activity_ocr_activity.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".OCR_activity">
+    tools:context="com.vitpunerobotics.netra.OCR_activity">
 
     <SurfaceView
         android:id="@+id/surface_camera_preview"
