+++ app/src/main/res/layout/activity_option.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context="com.vitpunerobotics.netra.MainActivity">
 
     <LinearLayout
         android:layout_width="368dp"
