+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -8,7 +8,7 @@
     android:layout_gravity="center"
     android:background="@color/trflogo_black"
     android:orientation="vertical"
-    tools:context=".splashScreen">
+    tools:context="com.vitpunerobotics.netra.splashScreen">
 
     <LinearLayout
         android:layout_width="match_parent"
