+++ app/src/main/res/layout/activity_voice__control.xml
@@ -8,7 +8,7 @@
     android:background="@color/yellow"
     android:orientation="horizontal"
     android:weightSum="3"
-    tools:context=".Voice_Control">
+    tools:context="com.vitpunerobotics.netra.Voice_Control">
 
     <TextView
         android:layout_width="wrap_content"
