+++ app/src/main/res/layout/content_main3.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Main3Activity">
+    tools:context="com.vitpunerobotics.netra.Main3Activity">
 
     <SurfaceView
         android:id="@+id/surface_camera_preview"
